Check commit

+++ marketplace-server/src/main/java/com/ncgroup/marketplaceserver/goods/model/Good.java
@@ -21,7 +21,7 @@ public class Good {
     private long id;
     private String goodName;
     private String firmName;
-    private int quantity;
+    private double quantity;
     private double price;
 
     private Unit unit;
@@ -35,6 +35,7 @@ public class Good {
     private String categoryName;
     private String image;
     private boolean status;
+
 
     public Good(GoodDto goodDto, Long id, MediaService mediaService) {
         this.setStatus(true);
